[Silver III] Title: 주유소, Time: 20 ms, Memory: 2800 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 int arr1[100001]; //주유소
 int arr2[100001]; //도로
-int ans = 0;
+long long ans = 0;
 int mini;
 int main(){
     ios::sync_with_stdio(false);
@@ -18,12 +18,12 @@ int main(){
         if(i ==0) mini = arr1[i];
     }
 
-    ans += arr1[0]*arr2[0];
+    ans += (long long)arr1[0]*arr2[0];
     for(int i = 1; i < N-1; i++){
         if(arr1[i] > mini){
-            ans += mini*arr2[i];
+            ans += (long long)mini*arr2[i];
         }else{
-            ans += arr2[i]*arr1[i];
+            ans += (long long)arr2[i]*arr1[i];
             mini = arr1[i];
         }
     }
